PP-1080 Address code review comments
