+++ src/main/java/uk/gov/pay/publicauth/resources/PublicAuthResource.java
@@ -7,7 +7,7 @@ import io.dropwizard.auth.Auth;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.publicauth.dao.AuthTokenDao;
-import uk.gov.pay.publicauth.model.TokenState;
+import uk.gov.pay.publicauth.model.TokenStateFilterParam;
 import uk.gov.pay.publicauth.model.Tokens;
 import uk.gov.pay.publicauth.service.TokenService;
 
@@ -25,7 +25,7 @@ import static java.util.Arrays.asList;
 import static java.util.UUID.randomUUID;
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static javax.ws.rs.core.Response.*;
-import static uk.gov.pay.publicauth.model.TokenState.ACTIVE;
+import static uk.gov.pay.publicauth.model.TokenStateFilterParam.ACTIVE;
 import static uk.gov.pay.publicauth.util.ResponseUtil.*;
 
 @Singleton
@@ -85,7 +85,7 @@ public class PublicAuthResource {
     @Path(FRONTEND_AUTH_PATH + "/{accountId}")
     @Produces(APPLICATION_JSON)
     @GET
-    public Response getIssuedTokensForAccount(@PathParam("accountId") String accountId, @QueryParam("state") TokenState state) {
+    public Response getIssuedTokensForAccount(@PathParam("accountId") String accountId, @QueryParam("state") TokenStateFilterParam state) {
         state = Optional.ofNullable(state).orElse(ACTIVE);
         List<Map<String, Object>> tokensWithoutNullRevoked = authDao.findTokensWithState(accountId, state);
         return ok(ImmutableMap.of("tokens", tokensWithoutNullRevoked)).build();
